+++ crates/navara_math/src/trigonometry.rs
@@ -8,7 +8,7 @@ pub fn chord_length<F: Float + Two<F>>(angle: F, radius: F) -> F {
 
 #[cfg(test)]
 mod test {
-    use std::f32::consts::PI;
+    use std::f64::consts::PI;
 
     use approx::assert_abs_diff_eq;
 
@@ -22,12 +22,12 @@ mod test {
         assert_abs_diff_eq!(chord_length(PI / 3., 5.), 5., epsilon = EPSILON6);
         assert_abs_diff_eq!(
             chord_length(PI / 3. * 2., 1.),
-            3.0f32.sqrt(),
+            3.0f64.sqrt(),
             epsilon = EPSILON6
         );
         assert_abs_diff_eq!(
             chord_length(PI / 3. * 2., 5.),
-            5.0 * 3.0f32.sqrt(),
+            5.0 * 3.0f64.sqrt(),
             epsilon = EPSILON6
         );
         assert_abs_diff_eq!(chord_length(PI, 10.), 2.0 * 10.0, epsilon = EPSILON6);
