+++ crates/navara_mvt/src/tile/traverse.rs
@@ -145,7 +145,7 @@ pub fn traverse_tile(
         globe.max_sse
     } else {
         vector_tile_appearance.max_sse
-    };
+    } as f64;
     let meets_sse = sse <= max_sse;
 
     let is_renderable = is_rendered_last_frame || is_tile_ready;
